refactor: implement ChatPromptInput component
