+++ src/components/chat/chat-ui.tsx
@@ -1,86 +1,26 @@
-import { useContextTracking } from '@/hooks/use-context-tracking'
 import { useIsMobile } from '@/hooks/use-mobile'
-import { trackEvent } from '@/lib/posthog'
 import { cn } from '@/lib/utils'
 import { AnimatePresence, motion } from 'framer-motion'
-import { useCallback, useRef, useState } from 'react'
-import { useNavigate } from 'react-router'
-import { ContextOverflowModal } from '../context-overflow-modal'
-import { ContextUsageIndicator } from '../context-usage-indicator'
-import { PromptInput } from '../ui/prompt-input'
+import { useCallback, useRef } from 'react'
 import { SuggestionButtons } from './suggestion-buttons'
 import { useChatScrollHandler } from '@/chats/use-chat-scroll-handler'
 import { useChatState } from '@/chats/chat-state-provider'
-import { useChatData } from '@/chats/chat-data-provider'
 import { ChatMessages } from './chat-messages'
+import { ChatPromptInput, type ChatPromptInputRef } from './chat-prompt-input'
 
 export default function ChatUI() {
-  const { chatThread, id: chatThreadId, models } = useChatData()
+  const { hasMessages } = useChatState()
 
-  const { handleModelChange, handleSendMessage, handleStop, hasMessages, isStreaming, messages, selectedModel } =
-    useChatState()
+  const { resetUserScroll, scrollContainerRef, scrollHandlers, scrollTargetRef, scrollToBottom } =
+    useChatScrollHandler()
 
-  const { resetUserScroll, scrollContainerRef, scrollHandlers, scrollTargetRef, scrollToBottom } = useChatScrollHandler(
-    {
-      hasMessages,
-      isStreaming,
-      messages,
-    },
-  )
-
-  const [input, setInput] = useState('')
-  const [showOverflowModal, setShowOverflowModal] = useState(false)
-  const formRef = useRef<HTMLFormElement>(null)
-  const navigate = useNavigate()
+  const chatPromptInputRef = useRef<ChatPromptInputRef>(null)
   const { isMobile, isReady } = useIsMobile()
 
-  const { usedTokens, maxTokens, isContextKnown, isOverflowing } = useContextTracking({
-    model: selectedModel,
-    chatThreadId,
-    currentInput: input,
-    onOverflow: () => setShowOverflowModal(true),
-  })
-
-  const handleSubmit = async () => {
-    // Prevent submitting while streaming or if input is empty
-    const textToSend = input.trim()
-    if (isStreaming || !textToSend) return
-
-    if (isOverflowing) {
-      setShowOverflowModal(true)
-      trackEvent('chat_send_prompt_overflow', {
-        model: selectedModel,
-        length: textToSend.length,
-        prompt_number: messages.length + 1,
-      })
-      return
-    }
-
-    // Clear the input immediately for responsive UX
-    setInput('')
-
-    await handleSendMessage(textToSend)
-
-    // Reset user scroll state and scroll to bottom when submitting a new message
-    resetUserScroll()
-    requestAnimationFrame(() => {
-      scrollToBottom()
-    })
-  }
-
   const handleSelectPrompt = useCallback((prompt: string) => {
-    setInput(prompt)
-    requestAnimationFrame(() => {
-      const textareaElement = formRef.current?.querySelector('textarea')
-      if (textareaElement) {
-        textareaElement.focus()
-      }
-    })
+    chatPromptInputRef.current?.setInput(prompt)
+    chatPromptInputRef.current?.focus()
   }, [])
-
-  const handleNewChat = async () => {
-    await navigate('/chats/new')
-  }
 
   if (!isReady) {
     return null
@@ -138,26 +78,10 @@ export default function ChatUI() {
                 duration: 0.25,
               }}
             >
-              <PromptInput
-                ref={formRef}
-                chatThread={chatThread}
-                value={input}
-                onChange={(value: string) => setInput(value)}
-                placeholder="Say something..."
-                models={models}
-                selectedModelId={selectedModel.id}
-                onModelChange={handleModelChange}
-                showSubmitButton
-                onSubmit={handleSubmit}
-                isLoading={isStreaming}
-                isStreaming={isStreaming}
-                onStop={handleStop}
-                autoFocus
-                submitOnEnter={!isStreaming}
-                className="flex flex-col gap-2 bg-secondary p-4 rounded-md w-full"
-                footerStartElements={
-                  isContextKnown && <ContextUsageIndicator usedTokens={usedTokens ?? 0} maxTokens={maxTokens ?? 0} />
-                }
+              <ChatPromptInput
+                handleResetUserScroll={resetUserScroll}
+                handleScrollToBottom={scrollToBottom}
+                ref={chatPromptInputRef}
               />
             </motion.div>
 
@@ -176,13 +100,6 @@ export default function ChatUI() {
             )}
           </motion.div>
         </motion.div>
-
-        <ContextOverflowModal
-          isOpen={showOverflowModal}
-          onClose={() => setShowOverflowModal(false)}
-          maxTokens={maxTokens ?? undefined}
-          onNewChat={handleNewChat}
-        />
       </div>
     </div>
   )
